feat: Add new tab for list of projects
On click of See All button in projects tab, a new container will appear
from the bottom of the screen while the current content fades out. This
will be space for a list of all projects. This container will also need
to include a header and return button to original content.