feat: Change back button and update nav buttons
Updates the 'go back' button for project list container to an arrow
button in a new position next to the title that is more friendly with
other screen sizes. Also updates the nav buttons to me more symmetrical
and replaces the 'linkedin' button with an 'about me' button with
features to be implemented in future.